rename DeckPart to CardList
