+++ tests/routers/utils.py
@@ -1,13 +1,13 @@
 from beanie import PydanticObjectId
 from scooze.card import OracleCard
-from scooze.deckpart import DeckPart
+from scooze.cardlist import CardList
 from scooze.models.card import CardModel
 
 
-async def dict_from_deckpart(deck_part: DeckPart[OracleCard]) -> dict[PydanticObjectId, int]:
+async def dict_from_cardlist(card_list: CardList[OracleCard]) -> dict[PydanticObjectId, int]:
     deck_dict = {}
 
-    for c, q in deck_part.cards.items():
+    for c, q in card_list.cards.items():
         card = await CardModel.find_one({"name": c.name})
         if card is None:
             continue
